Added Tabs component to update AreaCharts component
