+++ package.json
@@ -23,6 +23,7 @@
     "@heroui/switch": "2.2.15",
     "@heroui/system": "2.4.13",
     "@heroui/table": "^2.2.18",
+    "@heroui/tabs": "^2.2.16",
     "@heroui/theme": "2.4.13",
     "@matejmazur/react-katex": "^3.1.3",
     "@react-aria/ssr": "3.9.7",
